[Lab07] Working out issues with simpleLoan getter and setter methods

+++ Lab07/src/Loan.h
@@ -23,7 +23,7 @@ public:
     // Getters
     float getPrincipal();
     float getInterestRate();
-    int getLoanLength();
+    float getLoanLength();
     std::string getLoanType();
 
     // Setters
